+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to
 
 - Diagnostic Settings entities and relationships for Azure API Management
   Services
+- Diagnostic Settings entities and relationships for Azure CDN
 - Diagnostic Settings entities and relationships for Azure Event Grid Domain and
   Azure Event Grid Topics
 - Diagnostic Settings entities and relationships for Azure Batch Account